# App/blueprint/static/swagger.yaml
swagger: '2.0'
info:
  title: "Library API"
  description: "This is an API to manage a Library"
  version: "1.0.0"

host: "127.0.0.1:5000"

schemes:
  - "http"

consumes:
  - "application/json"

produces:
  - "application/json"

securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header

paths:
  /members/login:
    post:
      tags:
        - Members
      summary: "Login endpoint (generates token)"
      description: "Validates member credentials, generates and returns a member-specific token."
      parameters:
        - in: "body"
          name: "body"
          description: "Login Credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"

  /members: #endpoint
    post: #Post request
      tags:
        - Members
      summary: "Endpoint to create/register a new user."
      description: "Endpoint to create a new member in our Library system, emails must be unique."
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a new Member."
          required: true
          schema:
            $ref: "#/definitions/CreateMemberPayload"
      responses:
        201:
          description: "Successfull Member Creation"
          schema:
            $ref: "#/definitions/CreateMemberResponse"
          examples:
            application/json:
              id: 1
              name: John Doe
              DOB: "1900-01-01"
              email: johnd@email.com
              password: password123
            
    get:
      tags:
        - Members
      summary: "Returns all Members"
      description: "Endpoint to retrieve a list of all members."
      responses:
        200:
          description: "Retrieved Members Successfully"           
          schema:
            $ref: "#/definitions/AllMembers" 

    put: 
      tags:
        - Members
      summary: "Endpoint to update a user."
      description: "Endpoint to update a member, this is a token authenticated route, and the user must be logged in to update their account."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to update a Member."
          required: true
          schema:
            $ref: "#/definitions/UpdateMemberPayload"
      responses:
        200:
          description: "Successfully Updated member"
          schema:
            $ref: "#/definitions/UpdateMemberResponse"
          examples:
            application/json:
              id: 1
              name: Dohn Joe
              DOB: "1900-01-01"
              email: dohnj@email.com
              password: password123   

definitions:

  LoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"
    required:
      - email
      - password
    
  LoginResponse:
    type: "object"
    properties:
      token:
        type: "string"
      message:
        type: "string"
      status:
        type: "string"

  CreateMemberPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      DOB:
        type: "string"
        format: date
      password:
        type: "string"
      
    required:
      - name
      - email
      - DOB
      - password


  CreateMemberResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      DOB:
        type: "string"
      password:
        type: "string"

  AllMembers:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        name:
          type: "string"
        email:
          type: "string"
        DOB:
          type: "string"
        password:
          type: "string"

  UpdateMemberPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      DOB:
        type: "string"
        format: "date"
      password:
        type: "string"
      
    required:
      - name
      - email
      - DOB
      - password


  UpdateMemberResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      DOB:
        type: "string"
      password:
        type: "string"